requirements.txt
flask
pandas
scipy
statsmodels
yfinance
matplotlib
numpy
plotly
flask-mysqldb
mysql-connector-python
mysqlclient
ConfigParser
gunicorn
pytest
SQLAlchemy
pymysql
py-healthcheck
healthcheck